Move `publish` task getter
