+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/dart/task/Publish.kt
@@ -28,7 +28,7 @@ package io.spine.internal.gradle.dart.task
 
 import io.spine.internal.gradle.TaskName
 import io.spine.internal.gradle.base.assemble
-import io.spine.internal.gradle.java.publish.publish
+import io.spine.internal.gradle.publish.publish
 import io.spine.internal.gradle.named
 import io.spine.internal.gradle.register
 import org.gradle.api.tasks.Copy
